Move item related enums to itemenums.h. Apply clang-format.
