+++ apps/freeablo/faworld/itemenums.cpp
@@ -1,0 +1,1 @@
+#include "itemenums.h"
